[40] - Controllers name refactor
