+++ app/Infrastructure/Controllers/WalletBalanceController.php
@@ -2,18 +2,22 @@
 
 namespace App\Infrastructure\Controllers;
 
-use App\Application\Services\WalletCryptocurrenciesService;
+use App\Application\Exceptions\WalletNotFoundException;
+use App\Application\Services\WalletBalanceService;
 use App\Infrastructure\Validators\WalletIdValidator;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller as BaseController;
 
-class GetsWalletCryptocurrenciesController extends BaseController
+class WalletBalanceController extends BaseController
 {
+    /**
+     * @throws WalletNotFoundException
+     */
     public function __invoke(
         $walletId,
         WalletIdValidator $walletIdValidator,
-        WalletCryptocurrenciesService $walletCryptocurrenciesService
+        WalletBalanceService $walletBalanceService
     ): JsonResponse {
         if (!$walletIdValidator->validateWalletId($walletId)) {
             return response()->json(
@@ -25,8 +29,13 @@ class GetsWalletCryptocurrenciesController extends BaseController
             );
         }
 
-        $walletCryptocurrencies = $walletCryptocurrenciesService->getWalletCryptocurrencies($walletId);
+        $balance = $walletBalanceService->getsBalance($walletId);
 
-        return response()->json($walletCryptocurrencies, Response::HTTP_OK);
+        return response()->json(
+            [
+                'balance_usd' => $balance
+            ],
+            Response::HTTP_OK
+        );
     }
 }
